Annoying " method was called at an unexpected time " exception

+++ InetSpeedTesting/Datum.h
@@ -2,12 +2,12 @@ namespace InetSpeedTesting
 {
 	public ref class Datum sealed
 	{
-		uint8 TestNum;
+		uint16 TestNum;
 		float32 RTT;
 		float32 UploadBW;
 		float32 DownloadBW;
 	public:
-		Datum(float32, float32, float32, float32);
+		Datum(uint16, float32, float32, float32);
 	};
 }
